feat: provide mechanism for updating stream schema version and avoiding the need to manually reset the local cache.
Adds a new streamSchemaVersion config that, when we need to change it,
will automatically clear the local cache and load a new personal stream
with that version number, so that the old stream info is preserved but
the new version of the app has a clean slate instead of breaking by
trying to read the old streams.

Also adds parameters to the stream specifying the kind of stream and the
schema version, so that it's possible for the app to detect what version
and kind of stream something is just from the stream metadata.